+++ packages/universal-header/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.2-rc.3](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/universal-header@2.7.2-rc.2...@twreporter/universal-header@2.7.2-rc.3) (2023-10-06)
+
+**Note:** Version bump only for package @twreporter/universal-header
+
+
+
+
+
+## [2.7.2-rc.2](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/universal-header@2.7.2-rc.1...@twreporter/universal-header@2.7.2-rc.2) (2023-10-05)
+
+**Note:** Version bump only for package @twreporter/universal-header
+
+
+
+
+
+## [2.7.2-rc.1](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/universal-header@2.7.2-rc.0...@twreporter/universal-header@2.7.2-rc.1) (2023-10-01)
+
+**Note:** Version bump only for package @twreporter/universal-header
+
+
+
+
+
+## [2.7.2-rc.0](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/universal-header@2.7.1...@twreporter/universal-header@2.7.2-rc.0) (2023-09-28)
+
+**Note:** Version bump only for package @twreporter/universal-header
+
+
+
+
+
 ## [2.7.1](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/universal-header@2.7.1-rc.4...@twreporter/universal-header@2.7.1) (2023-09-20)
 
 **Note:** Version bump only for package @twreporter/universal-header